Use init function for initializing Minim
